Working camera and TFT Eye
