+++ config/neopixel.yml
@@ -1,5 +1,5 @@
 neopixel:
-  enabled: true
+  enabled: false
   path: 'modules.neopixel.neopx.NeoPx'
   config:
     pin: 12 # Only used for GPIO. GPIO2 and 3 are use for i2c, GPIO10 is used for SPI, GPIO12 is used for GPIO
