+++ config/piservo.yml
@@ -4,9 +4,10 @@ piservo:
   instances:
     - name: "ear"
       id: 0
-      pin: 12 
+      pin: 22
       range: [-40, 40]
       start: 0
+      test_on_boot: false
   dependencies:
     python:
       - gpiozero
