+++ config/tft_display.yml
@@ -1,13 +1,23 @@
 tft_display:
-  enabled: false
-  path: modules.display.tft_display.TFTDisplay
+  enabled: true
+  path: modules.display.tft_display_eye.TFTDisplayEye
   config:
     bus: 0 # Use `ls /dev/spi*` to identify. (/dev/spidev0.0) is bus 0, device 0
     device: 0 # Use `ls /dev/spi*` to identify. (/dev/spidev0.0) is bus 0, device 0
     rst_pin: 27
     dc_pin: 25
     bl_pin: 18 # Not used if no backlight pin on device
+    rotation: -90.0
     test_on_boot: true
+    colors:
+      red: (255, 32, 32)
+      blue: (0, 245, 255)
+      green: (0, 255, 180)
+      purple: (255, 0, 255)
+      yellow: (255, 255, 0)
+      white: (255, 255, 255)
+      gray: (128, 128, 128)
+      dark_gray: (64, 64, 64)
   dependencies:
     unix:
       - python3-pip
