+++ config/tracking.yml
@@ -1,5 +1,5 @@
 tracking:
-  enabled: true
+  enabled: false
   path: 'modules.vision.imx500.tracking.Tracking'
   config:
     active: true
