+++ modules/display/tft_display_eye.py
@@ -1,0 +1,97 @@
+from PIL import Image, ImageDraw
+from modules.display.tft_display import TFTDisplay
+import time
+
+class TFTDisplayEye(TFTDisplay):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        # Ensure colors are stored as tuples
+        self.colors = {
+            key: tuple(map(int, value.strip('()').split(',')))
+            for key, value in kwargs.get('colors', {}).items()
+        }
+        if kwargs.get('test_on_boot'):
+            self.init_eye()
+
+    def setup_messaging(self):
+        self.subscribe('eye', self.eye)
+
+    def init_eye(self):
+        self.draw_image('makerforge_bl.png')
+        time.sleep(2)
+        img = None
+        
+        # Increase radius from 0 to 70 in steps of 5
+        for x in range(0, 70, 5):
+            img = self.draw_halo(
+                ring_radius=x,
+                color=self.colors['blue'],
+                glow_spread=10
+            )
+        # increase glow spread from 10 to 30 in steps of 5
+        for x in range(10, 30, 2):
+            img = self.draw_halo(
+                color=self.colors['blue'],
+                glow_spread=x
+            )
+        img = self.eye('blue')
+        time.sleep(3)
+        self.blink(img)
+        print("TFT display test completed")
+
+    def eye(self, color):
+        # print(f"Eye color: {color}")
+        # If color doesn't exist, throw exception
+        if color not in self.colors:
+            raise ValueError(f"Color '{color}' not found in available colors.")
+        # Access the color as a tuple
+        return self.draw_halo(
+            ring_radius=70,
+            ring_thickness=10,
+            color=self.colors[color]
+        )
+
+    def draw_halo(self, color, ring_radius=70, ring_thickness=10, glow_layers=12, glow_spread=30):
+        disp = self.disp
+        center = (disp.width // 2, disp.width // 2)
+
+        glow_image = Image.new("RGBA", (disp.width, disp.height), (0, 0, 0, 0))
+        draw = ImageDraw.Draw(glow_image)
+
+        for i in range(glow_layers):
+            spread = int(i * glow_spread / glow_layers)
+            alpha = int(255 * (1 - i / glow_layers) ** 2 * 0.5)
+            glow_color = (color[0], color[1], color[2], alpha)
+
+            outer_radius = ring_radius + ring_thickness // 2 + spread
+            inner_radius = ring_radius - ring_thickness // 2 - spread
+
+            outer_bbox = [
+                center[0] - outer_radius,
+                center[1] - outer_radius,
+                center[0] + outer_radius,
+                center[1] + outer_radius
+            ]
+            draw.ellipse(outer_bbox, outline=glow_color, width=3)
+
+            if inner_radius > 0:
+                inner_bbox = [
+                    center[0] - inner_radius,
+                    center[1] - inner_radius,
+                    center[0] + inner_radius,
+                    center[1] + inner_radius
+                ]
+                draw.ellipse(inner_bbox, outline=glow_color, width=3)
+
+        ring_bbox = [
+            center[0] - ring_radius - 2,
+            center[1] - ring_radius - 2,
+            center[0] + ring_radius + 2,
+            center[1] + ring_radius + 2
+        ]
+        draw.ellipse(ring_bbox, outline=(color[0], color[1], color[2], 255), width=ring_thickness)
+
+        final_image = Image.new("RGB", (disp.width, disp.height), "BLACK")
+        final_image.paste(glow_image, (0, 0), glow_image)
+        disp.ShowImage(final_image)
+        return final_image
